+++ src/main/java/duke/Deadline.java
@@ -1,5 +1,6 @@
+package duke;
+
 import java.time.LocalDate;
-import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
 
 public class Deadline extends Task {
